+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Process behaviour.
+- Output textfield cannot be copied when empty.
 
 ## [0.8.0] - 2025-03-26